requirements.txt
langchain
langchain_community
streamlit
langchain_chains
faiss
python-dotenv